+++ ExtendedCheats/Commands.cs
@@ -1,7 +1,8 @@
-﻿using BepInEx;
-using BepInEx.Configuration;
+﻿using BepInEx.Configuration;
+using HarmonyLib;
 using System;
-using System.IO;
+using System.Collections;
+using System.Reflection;
 using UnityEngine;
 
 namespace Desperados3Mods.ExtendedCheats
@@ -9,6 +10,7 @@ namespace Desperados3Mods.ExtendedCheats
     static class Commands
     {
         static bool show = false;
+        static readonly FieldInfo fieldMissionSetupSettings_s_difficultySettings = AccessTools.Field(typeof(MissionSetupSettings), "s_difficultySettings");
 
         public static void Bind(ConfigFile config)
         {
@@ -65,7 +67,7 @@ namespace Desperados3Mods.ExtendedCheats
 
                     var ammoType = GetAmmoType(playerSkill);
                     if (ammoType == 0) continue;
-                    
+
                     var charSkill = new CharacterSkillName(skill.m_skillData.name);
 
                     DrawAdjustableInt(charSkill.skill + " Ammo", $"{playerSkill.iCount}/{character.m_charInventory.MaxCount(ammoType)}",
@@ -86,6 +88,53 @@ namespace Desperados3Mods.ExtendedCheats
                 File.WriteAllText(Path.Combine(Paths.ConfigPath, "skills.toml"), content);
             }
 #endif
+            var difficultySettings = (MissionSetupSettings.DifficultySettings)fieldMissionSetupSettings_s_difficultySettings.GetValue(null);
+            var showdownPauseOld = difficultySettings.m_bFocusModePause;
+            difficultySettings.m_bFocusModePause = GUILayout.Toggle(difficultySettings.m_bFocusModePause, "Pausing Showdownmode");
+            if (showdownPauseOld != difficultySettings.m_bFocusModePause)
+            {
+                fieldMissionSetupSettings_s_difficultySettings.SetValue(null, difficultySettings);
+            }
+
+            if (SceneStatistics.instance != null && GUILayout.Button("Reset Level Statistics"))
+            {
+                var stats = SceneStatistics.instance;
+                var statsType = typeof(SceneStatistics);
+                ((IDictionary)AccessTools.Field(statsType, "m_dictNPCsKilledBy").GetValue(stats)).Clear();
+                ((IDictionary)AccessTools.Field(statsType, "m_dictSkillsUsed").GetValue(stats)).Clear();
+                AccessTools.Field(statsType, "m_iAlarmCount").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iBodiesHidden").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iCritterKilled").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iDeadBodiesFound").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iEnemiesKilled").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iEnemiesKilledWithTorches").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iEnemiesShotPlayer").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iDeadBodiesFound").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iEnemiesSpared").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iEnemiesTiedUp").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iDeadBodiesFound").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iNPCsKilled").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iPlayerDetected").SetValue(stats, 0);
+                AccessTools.Field(statsType, "m_iSamuraiKilled").SetValue(stats, 0);
+
+                var statsData = stats.persistentSaveData;
+                statsData.fPlaytime = 0;
+                statsData.iFrameCount = 0;
+                statsData.iLoadCount = 0;
+                statsData.iSaveCount = 0;
+                statsData.iTotalHealthLost = 0;
+                statsData.lPlayerDeathCount.Clear();
+                statsData.lPlayerStateDurations.Clear();
+
+                var statsNoSaveData = SceneStatisticsNoSave.instance.persistentSaveData;
+                statsNoSaveData.fPlaytime = 0;
+                statsNoSaveData.iFrameCount = 0;
+                statsNoSaveData.iLoadCount = 0;
+                statsNoSaveData.iSaveCount = 0;
+                statsNoSaveData.iTotalHealthLost = 0;
+                statsNoSaveData.lPlayerDeathCount.Clear();
+                statsNoSaveData.lPlayerStateDurations.Clear();
+            }
 
             GUILayout.EndVertical();
             GUILayout.EndVertical();